requirements.txt
textual
python-chess 